hashseg: Add option to generate MBN header v6
MBN header v6 seems to be used in SM8150. It looks like the metadata
(e.g. SW_ID/HW_ID etc) is moved from the certificate chain to an extra
part in the hash segment (see also the v2.0 documentation [1]).

It's not clear yet how the metadata structure looks, so just fill it
with zeroes for now and add the necessary MBN header changes (adapted
from coreboot). It looks like the hash digest is also changed from
SHA256 to SHA384 for MBN v6.

[1]: https://www.qualcomm.com/media/documents/files/secure-boot-and-image-authentication-technical-overview-v2-0.pdf