hashseg: Fix swapped image_id/MBN header version definition
It looks like the definition of the MBN header in signlk is wrong.
The 0x3 actually refers to the MBN header version and the first field
is the image_id (which seems to be sometimes set but overall unused).

Correct this with a new reference to coreboot, which seems to be a bit
less messy than signlk.