hashseg: Add option to generate MBN header v5
MBN header v5 seems to be used in SDM845. It is not much different,
according to the coreboot source code some of the header fields seem
to be reused to refer to a second signature from Qualcomm (rather than
the OEM).

Add some abstractions and a -v5 option to generate the newer header
instead.