hashseg: Fix MBNv7 total_size and clarify descriptions
At the moment, we model the "common metadata" as part of the MBNv7 header,
so the common_metadata_size is part of the header size (included in the
FORMAT.size). This means we should not add it to the "total_size", or it
ends up in there twice.

Remove it from total_size, add an assertion to ensure the total_size always
matches what we actually produce, and clarify the descriptions a bit to
make that more clear.