hashseg: Set correct hash_table_size for v7
"hash_table_size" in v7 seems to be essentially the "hash_size" in older
MBN header versions. We should set it to the total size of all hashes we
include in the header. The "_HashSegment" base class does that already when
the field is called "hash_size", so let's just rename it to make that work.