hashseg: Clarify ELF PHDR flag definition based on Coreboot
Looks like the flags in the ELF PHDR used by QC represent certain fields
that we can access with masks and shifts. The access type used for the hash
segment on newer platforms seems to have changed, so we no longer manage to
drop the hash segment in that case.

Fix this by checking just the relevant bits and use the chance to clarify
the bits we use.